requirements.txt
pillow==11.1.0
pygame==2.6.1
pygbag==0.9.2
